First version with SML, getting power sign from meter. Note that LCD RS is now GPIO16 !
Code is now receiving the SML data stream via a second optical input from power meter. Currently the "reduced dataset" is only supported, which is already available from the meter without a PIN code. It does not contain the current power, but it contains the "meter status word", which contains the direction sign, if we currently consume or feed power into the grid. In combination with the absolute power value from the optical impulse, we have the complete power value now. However, the value is currently only displayed on the LCD. The power sign is not used anywhere in the code yet.

The LCD RS pin GPIO has been changed from 12 to 16, because GPIO 12 needs to be high during programming. That was not always the case with the LCD connected, so I got a programming error every second time. This error is now gone with GPIO 16.

A long button press is now supported. It stops the code and puts it into a "programming mode". This way faulty output towards the Multiplus is avoided during programming. To also avoid that the Multiplus turns off during programming & reset of the ESP32, add a 1kOhm resistor from Gnd to GPIO27 = VEBUS_DE.