immediate ESS calculation without settling using Pmains (ACin power)
Settling has been completely removed. Instead the ESS compensation power for the other AC phases is now calculated from the electric meter power plus the actual/real power on Multiplus ACin. This power is polled 10 times per second from the multiplus. From this power, an powerACin value is calculated, which is in sync with last electric meter value.
Benefit is, that now also the compensation of power in front or in parallel of the multiplus is very fast and precise, as we react and re-adjust the ESS power every second.

Additioally, a first attempt to turn on a Shelly is implemented, in case the battery is full and there is excessive power to spend.

Note that as the code is now stripped from settling, the old ESS power calculation routines for the absolute power meter are currently not supported and will currently not function.